Make symtree work with 64-bit Windows binaries.

+++ find_refs.h
@@ -4,7 +4,7 @@
 #include <functional>
 #include <vector>
 
-void find_refs_x86(size_t addr, bool x64, std::vector<uint8_t> const & buf, std::function<void (size_t addr)> const & cb);
-void find_refs_arm32(size_t addr, std::vector<uint8_t> const & buf, std::function<void (size_t addr)> const & cb);
+void find_refs_x86(uint64_t addr, bool x64, std::vector<uint8_t> const & buf, std::function<void (uint64_t addr)> const & cb);
+void find_refs_arm32(uint64_t addr, std::vector<uint8_t> const & buf, std::function<void (uint64_t addr)> const & cb);
 
 #endif // SYMTREE_FIND_REFS_H
